销售人员接口完成

+++ obd-core/src/main/java/com/wkhmedical/service/impl/BindUserServiceImpl.java
@@ -11,12 +11,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.taoxeo.lang.BeanUtils;
 import com.taoxeo.lang.exception.BizRuntimeException;
-import com.wkhmedical.dto.BindUserBody;
+import com.wkhmedical.dto.BindUserAddBody;
 import com.wkhmedical.dto.BindUserDTO;
+import com.wkhmedical.dto.BindUserEditBody;
 import com.wkhmedical.dto.BindUserPage;
 import com.wkhmedical.dto.BindUserParam;
 import com.wkhmedical.po.BindUser;
 import com.wkhmedical.repository.jpa.BindUserRepository;
+import com.wkhmedical.repository.jpa.CarInfoRepository;
 import com.wkhmedical.service.BindUserService;
 import com.wkhmedical.util.AssistUtil;
 import com.wkhmedical.util.BizUtil;
@@ -29,6 +31,8 @@ public class BindUserServiceImpl implements BindUserService {
 
 	@Resource
 	BindUserRepository bindUserRepository;
+	@Resource
+	CarInfoRepository carInfoRepository;
 
 	@Override
 	public BindUser getInfo(BindUserParam paramBody) {
@@ -51,10 +55,26 @@ public class BindUserServiceImpl implements BindUserService {
 	}
 
 	@Override
-	public void addInfo(BindUserBody infoBody) {
+	@Transactional
+	public void addInfo(BindUserAddBody infoBody) {
+		Long id = BizUtil.genDbId();
+		// 校验默认
+		int isDefault = infoBody.getIsDefault();
+		int utype = infoBody.getUtype();
+		int areaId = infoBody.getAreaId();
+		if (isDefault == 1) {
+			BindUser buTmp = bindUserRepository.findByUtypeAndAreaIdAndIsDefault(utype, areaId, isDefault);
+			if (buTmp != null) {
+				// 已存在默认人员设置，则不允许添加默认人员
+				throw new BizRuntimeException("binduser_already_default");
+			}
+			// 新增默认人员时，将相应区县车辆与相关人员对应
+			carInfoRepository.updateCarInfoBindUser(id, utype, areaId);
+		}
+
 		// 组装Bean
 		BindUser bindUser = AssistUtil.coverBean(infoBody, BindUser.class);
-		bindUser.setId(BizUtil.genDbId());
+		bindUser.setId(id);
 		bindUser.setDelFlag(0);
 		// 入库
 		bindUserRepository.save(bindUser);
@@ -62,7 +82,7 @@ public class BindUserServiceImpl implements BindUserService {
 
 	@Override
 	@Transactional
-	public void updateInfo(BindUserBody infoBody) {
+	public void updateInfo(BindUserEditBody infoBody) {
 		// id检查
 		Long id = infoBody.getId();
 		if (id == null) {
@@ -74,6 +94,32 @@ public class BindUserServiceImpl implements BindUserService {
 			throw new BizRuntimeException("info_not_exists", id);
 		}
 		BindUser bindUserUpd = optObj.get();
+		// 处理人员分类更改的情况
+		Integer utype = infoBody.getUtype();
+		if (utype != null) {
+			if (!utype.equals(bindUserUpd.getUtype())) {
+				// 更新人员分类时，则需将原对应车辆设置为空
+				carInfoRepository.updateCarInfoBindUserNull(id, bindUserUpd.getUtype());
+			}
+		}
+		else {
+			utype = bindUserUpd.getUtype();
+		}
+		// 校验默认
+		Integer isDefault = infoBody.getIsDefault();
+		if (isDefault != null) {
+			// 人员更改设置为默认的情况
+			if (isDefault.intValue() == 1) {
+				// 校验是否已存在默认人员
+				BindUser buTmp = bindUserRepository.findByUtypeAndAreaIdAndIsDefault(utype, bindUserUpd.getAreaId(), isDefault);
+				if (buTmp != null && !id.equals(buTmp.getId())) {
+					// 已存在默认人员设置，则不允许更改默认人员
+					throw new BizRuntimeException("binduser_already_default");
+				}
+				// 将相应区县车辆与相关人员对应
+				carInfoRepository.updateCarInfoBindUser(id, utype, bindUserUpd.getAreaId());
+			}
+		}
 		// merge修改body与原记录对象
 		BeanUtils.merageProperty(bindUserUpd, infoBody);
 		// 更新库记录
@@ -82,7 +128,15 @@ public class BindUserServiceImpl implements BindUserService {
 
 	@Override
 	public void deleteInfo(Long id) {
-		bindUserRepository.deleteById(id);
+		Optional<BindUser> optObj = bindUserRepository.findById(id);
+		if (optObj.isPresent()) {
+			BindUser bindUserUpd = optObj.get();
+			// 设置相应车辆负责人为空
+			carInfoRepository.updateCarInfoBindUserNull(id, bindUserUpd.getUtype());
+			// 删除
+			bindUserRepository.deleteById(id);
+		}
+
 	}
 
 	@Override
